refactor(card): extract evolution writer
