+++ lib/core/services/card_rewriter/card_evolution_collector_coordinator.dart
@@ -628,9 +628,6 @@ class CardEvolutionCollectorCoordinator {
     }
   }
 
-  Map<String, String>? retrievalTargets(String selectedInputJson) =>
-      _retrievalTargets(selectedInputJson);
-
   void cancelSession(String sessionId) {
     _observationTokens[sessionId]?.cancel('generationAborted');
   }
